pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-model-rag"
dynamic = ["version"]
authors = [
    {name = "Zirui Guo"}
]
description = "Multi-Model-RAG: All-in-One RAG System"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "huggingface_hub",
    "lightrag-hku",
    "mineru[core]",
    "tqdm",
]

[project.optional-dependencies]
image = ["Pillow>=10.0.0"]
text = ["reportlab>=4.0.0"]
office = []  # Requires LibreOffice (external program)
paddleocr = [
    "paddleocr>=2.7.0",
    "pypdfium2>=4.25.0",
]
markdown = [
    "markdown>=3.4.0",
    "weasyprint>=60.0",
    "pygments>=2.10.0",
]
dashboard = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "ollama>=0.4.0",
]
quality = [
    "FlagEmbedding>=1.3.5",
    "instructor>=1.7.0",
    "ragas>=0.2.0",
]
all = [
    "Pillow>=10.0.0",
    "reportlab>=4.0.0",
    "paddleocr>=2.7.0",
    "pypdfium2>=4.25.0",
    "markdown>=3.4.0",
    "weasyprint>=60.0",
    "pygments>=2.10.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "ollama>=0.4.0",
    "FlagEmbedding>=1.3.5",
    "instructor>=1.7.0",
    "ragas>=0.2.0",
]

[project.urls]
Homepage = "https://github.com/HKUDS/Multi-Model-RAG"
Documentation = "https://github.com/HKUDS/Multi-Model-RAG"
Repository = "https://github.com/HKUDS/Multi-Model-RAG"
Issues = "https://github.com/HKUDS/Multi-Model-RAG/issues"

[tool.uv]
dev-dependencies = [
    "pytest>=6.0",
    "pytest-asyncio",
    "black",
    "isort",
    "flake8",
    "mypy",
    "openai",
    "python-dotenv",
]

[tool.setuptools.packages.find]
include = ["multi_model_rag*"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
multi_model_rag = ["py.typed"]

[tool.setuptools.dynamic]
version = {attr = "multi_model_rag.__version__"}

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
# Pragmatic starter set: pyflakes + pycodestyle-errors, isort,
# flake8-bugbear, flake8-simplify, Ruff-native rules.
# Opinionated stylistic rules (UP = pyupgrade, PTH = pathlib) are deliberately
# left out of the default run — enable them module-by-module as code is
# modernised to avoid a large churn-only commit.
select = ["E", "F", "I", "B", "SIM", "RUF"]
ignore = [
    "E402",     # module-level import not at top (load_dotenv pattern)
    "E501",     # line length (ruff-format handles what it can)
    "E722",     # bare except (existing code relies on it; tighten gradually)
    "B008",     # function call in default argument (common FastAPI pattern)
    "B904",     # raise ... from err (pending broader exception cleanup)
    "SIM102",   # nested if collapsing (often hurts readability)
    "SIM108",   # ternary expressions (readability subjective)
    "SIM117",   # combining with statements
    "RUF001",   # ambiguous unicode in docstrings (zh-CN prompts trip this)
    "RUF002",   # ambiguous unicode in docstrings
    "RUF003",   # ambiguous unicode in comments
    "RUF012",   # mutable class attrs should use ClassVar
    "RUF013",   # implicit Optional (mass cleanup needed)
    "RUF005",   # iterable unpacking vs concatenation (style)
    "SIM105",   # contextlib.suppress vs try/except/pass (style)
    "SIM118",   # `key in dict` vs `key in dict.keys()` (minor style)
    "B007",     # unused loop control variable (intentional in some cases)
]

[tool.ruff.lint.per-file-ignores]
"examples/**" = ["E402", "F401", "F841", "B"]
"tests/**" = ["F401", "F811", "B"]
"multi_model_rag/prompt*.py" = ["E501"]
"multi_model_rag/prompts_zh.py" = ["E501", "RUF001", "RUF002", "RUF003"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
check_untyped_defs = true
files = ["multi_model_rag"]
exclude = ["build/", "dist/", "examples/", "tests/"]

[[tool.mypy.overrides]]
# Start lenient; tighten module-by-module over time.
module = "multi_model_rag.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test*.py"]
norecursedirs = ["examples"]
